Impl for getProducts service.
* Testing getProduct service.

* Implemented proper service errors.

* GetProductService params updated.

* GetProducts.kt handled case for no products available.

* Changed all id type to long from int.

* changed uid parsing from int to long in GetProducts.kt route.

* Updated user data in Mapper.kt for id type change.

* Added category id to Product classes.

+++ src/main/kotlin/onlytrade/app/utils/ImageUploadService.kt
@@ -20,10 +20,10 @@ object ImageUploadService {
 
     fun buildImagePath(
         rootFolderName: String,
-        userId: Int,
-        categoryId: Int,
-        subcategoryId: Int,
-        productId: Int
+        userId: Long,
+        categoryId: Long,
+        subcategoryId: Long,
+        productId: Long
     ) = "$rootFolderName/$userId/$categoryId/$subcategoryId/$productId"
 
     fun uploadFile(
